Provide a better exception when a PrimitiveValue is encountered when trying to read a ComplexType
Currently, when we parse the JSON for a ComplexType property, if we
encounter a PrimitiveValue, we get back an exception that is mostly
useless to a client:

An unexpected 'PrimitiveValue' node was found when reading from the JSON
reader. A 'StartObject' node was expected.

This change makes the exception provide the property name and a
description of what was expected, including the type name.