Update to make sure all valid odata annotations parsed by an ODataMessageReader show up in the resulting ODataEntry
Prior to this change, any valid custom annotations present in a request
body would show up in the "InstanceAnnotations" of an entity, complex
type, or property, but any odata annotations (those starting with
"odata." or, with simplified parsing, no namespace) will not show up.
They will direct the parser in how to handle the data, but it is not
possible to use them in code.

This change makes sure the valid annotations (known odata ones) show up
in the InstanceAnnotations.