Fix for https://github.com/OData/odata.net/issues/592
Description of problem:
Suppose there is a function getCurrentCustomer() whose EntitySetPath is not defined, if the request URL is /getCurrentCustomer()/oneNavigationProperty, the ODataPathParser.Parse() will throw NullReferenceException as the current code requires EntitySetPath to be specified on Function that returns an enitty type.

Description of fix:
As the EdmNavigationSource is not required for NavigationPropertySegment, we do not need to require the TargetEdmNavigationSource.