Adding .NET Standard libraries to E2E test. Everything builds!
Caveat: there's a bug in VIsual Studio that causes the xproj to have
trouble finding the csproj files for ODL. Therefore committing the
project.lock.json file as well so that we don't have to run nuget
restore.

Also adjusting the build script for binary output to address the E2E
xproj build issue.