This change fixes issue #789.
The EdmModel constructor was failing with System.TypeInitializationException:“The type initializer
for 'Microsoft.OData.Edm.Vocabularies.V1.CoreVocabularyModel' threw an exception.”

This was caused by the construction of the CoreVocabularyModel looking for embedded resources
that are not present in the .NetStandard version of Edm.

In addition to adding the embedded resources, the resource name are fully qualified
in NetStandard so addition code is added to find the proper name for the resource before
loading it.

The fix was tested by temporarily adding a unit test to validate the EdmModel constructor
as well as the unit test validation for CoreVocabularyModel and AlternateKeysVocabularyModel.
These tests are not included in this change as we need to invest in more than just these tests.
A future change will introduce a test suite for the NetStandard binary.