This change fixes test failures in both local machine and in the lab because of the switch to VS2015. Switching to VS2015 changed the testing database engine from SQL Express to LocalDb. Since VS2015 installs LocalDb, this is a change that make sense to keep.
For local tests, some client tests were failing due to a failure to
clean up local database entries in LocalDB. The change to artdbclean.js
uses SQL Native Client for LocalDB to fix a connection issue, allowing the
script to properly clean up after the tests.

For lab tests, the machine configuration is a bit different and LocalDb is not
available; therefore, use SQL Express when building the lab. Since the current
VS2015 symbol was only used for test DB selection, I renamed the symbol to
reflect its usage.

- [   ] Test cases added
- [ x ] Build and test with one-click build and test script passed