Updating build script for Windows Store test solution
Updating build.ps1 to skip building the WindowsStore test solution if VS2013 is not installed. This is due to the fact that Windows 8.0 apps are not supported with VS2015.

Also updating the solution file to build the projects in order in the event of race conditions during build.

And adding .NET Core packages to nuget for the .NET Standard solution.