Migrating .NET Core tests and .NET Standard source projects to VS2017
- OData.Tests.NetStandard.sln is now OData.Tests.NetStandard.VS2017.sln
to indicate that it is to be opened with VS2017. Because we have a
massive number of projects to migrate from VS2015, we will do them
incrementally and use the VS2017 suffix to indicate which solutions are
VS2017. Once we are done with the migration, we'll remove the suffix.
- Replaced .xproj with new VS2017 .csproj for .NET Core tests. You will
need to have VS2017 installed if you wish to run the unit tests in .NET
Core. As a result of the migration, also cleaning up the project.jsons
and related files.
- Added VS2017 .csproj versions of the .NET Standard product source. You
may still launch OData.NetStandard.sln to modify and build .NET Standard
with VS2015. However, once we are in a state where we can flip the
switch to be 100% VS2017, OData.NetStandard.sln will also be in VS2017.
- Deleted src\Microsoft.OData.Core\PortabilityExtensionMethods.cs as it
is not used
- Fixing a few stylecop errors
- Updating build.ps1 to build the .NET Core tests if VS2017 is installed