Adding new Microsoft.Test.OData.Tests.Client.NetCore
- Adding new tests for OData Client in NetCore. These tests currently do
not run properly as they need to talk to a test service. The NetFx
version of these tests are running the test service in-proc whereas
these tests are talking to the service out-proc. The next step is to
launch the out-proc service to have the tests query against it.
- The test project currently compiles with stubs for the interface that
will be used to talk to the service.