This change adds nuget package testing.
This intention is to provide a small test project that can ingest the latest NuPkg's
from the build, restore them, build the project which results in a testable Dll. These
tests are very limited in scope, i.e. they only test the Dlls ability to load and
perform basic functions; the functional tests will test the full functionality.