Fixing lexer bug that breaks functions named 'in'
Judging by the behavior of all previous versions of OData, it should be legal
to have a custom function with the same name as an operator. However, the
introduction of the 'in' operator broke this for functions named 'in'. This
change ensures that functions named 'in' are treated the same as any other
function by the ExpressionLexer.