Update ODataTypeInfo.cs
I found an issue where the Conventional Id fields are not recognized as key fields.  As the Property Name Ends with Id, rather than "ID".  The result is that the EdmModel is properly defining the key property, however on the client the EntityType is resolved as "ComplexType" and failures in tracking occur later on in the pipeline.